integrate
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/perforce/cortical_surface/main@27638 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/library/surfacereferential/corticalConstraints.cc
@@ -639,8 +639,8 @@ TimeTexture<float> aims::EconstraintCleaner(TimeTexture<float> & texture, std::v
 	
 		//On passe un coup de PCC pour reperer les extremites les plus eloignees
 	
-	int path_max=0;
-	int path_temp=0;
+	float path_max=0;
+	float path_temp=0;
 	int ext1=0, ext2=0;
 	TimeTexture<float> tex_path(1,size);
 	TimeTexture<float> tex_path_temp(1,size);
